backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2
alembic==1.13.3
python-multipart==0.0.12
pydantic[email]
